Moved accommodation creation/update into AccommodationForm
Moved all responsibility for creating/editing an Accommodation to a component form; implement the form in the new accommodation modal
